129: address review — break sessions/ submodule cycle (Option E)
PR #150 review flagged a 3-cycle between the sessions submodules:

- lifecycle → detail   (SessionDetail, get_session_detail for create_session)
- detail   → races     (get_pending_races for SessionDetail.your_pending)
- races    → lifecycle (REJOIN_GRACE_MINUTES for the grace-window query)

Rust handles this fine — modules don't have init-order issues — but it
left the three concerns coupled. Resolves by hoisting the cross-cutting
items into a new sibling at the bottom of the dependency stack:

- sessions/types.rs (new): SessionRaceInfo, RaceSubmission, and
  REJOIN_GRACE_MINUTES. Re-exported from the module root, so external
  callers still reach them at crate::services::sessions::<name>.
- lifecycle.rs: no longer defines REJOIN_GRACE_MINUTES; imports from
  super::types. Still depends on super::detail for create_session's
  SessionDetail return type.
- detail.rs: no longer defines SessionRaceInfo / RaceSubmission;
  imports from super::types. Still depends on super::races for
  get_pending_races. RaceInfo and ParticipantInfo stay here (only
  SessionDetail consumes them).
- races.rs: imports SessionRaceInfo and REJOIN_GRACE_MINUTES from
  super::types instead of super::detail / super::lifecycle. No sibling
  imports remain.

Resulting graph (DAG): types ← {lifecycle, detail, races}, plus a
linear lifecycle → detail → races chain.

All 253 lib tests + 27 integration tests pass unchanged. clippy and
nightly fmt are clean.
