+++ backend/src/services/sessions.rs
@@ -1,9 +1,12 @@
 //! Session services — split by concern.
 //!
+//! - [`types`]: shared DTOs and constants (`SessionRaceInfo`,
+//!   `RaceSubmission`, `REJOIN_GRACE_MINUTES`). Lives below the other
+//!   submodules so the dependency graph stays acyclic.
 //! - [`lifecycle`]: create, join, leave, host transfer, close-stale, listing.
 //! - [`detail`]: aggregated read for the polling endpoint
-//!   (`get_session_detail`, `list_races`) plus the shared `SessionRaceInfo` /
-//!   `RaceSubmission` / `RaceInfo` / `ParticipantInfo` / `SessionDetail` DTOs.
+//!   (`get_session_detail`, `list_races`) plus its read-only DTOs
+//!   (`SessionDetail`, `ParticipantInfo`, `RaceInfo`).
 //! - [`races`]: race orchestration (next/skip track, pending races,
 //!   skip-pending).
 //!
@@ -13,7 +16,9 @@
 mod detail;
 mod lifecycle;
 mod races;
+mod types;
 
 pub use detail::*;
 pub use lifecycle::*;
 pub use races::*;
+pub use types::*;
